handle with long numbers

+++ include/swap.h
@@ -11,16 +11,16 @@
 #include <stdlib.h>
 
 int my_strlen(int const *str);
-int my_getnbr(char const *str);
-void swap(int *array);
-void swap_elem(int *array, int index1, int index2);
-void rotate_left(int *array, int size);
-void rotate_right(int *array, int size);
-int returnmin(int *array, int size);
-void putminfirstleft(int *array, int size, int min);
-void putminfirstright(int *array, int size, int min);
-int *fillsrc(int *src, int ac, char **av);
-void pushswap(int *la, int *lb, int *b, int size);
-void putinb(int *la, int *lb, int *b, int size);
+long my_getnbr(char const *str);
+void swap(long *array);
+void swap_elem(long *array, int index1, int index2);
+void rotate_left(long *array, int size);
+void rotate_right(long *array, int size);
+int returnmin(long *array, int size);
+void putminfirstleft(long *array, int size, long min);
+void putminfirstright(long *array, int size, long min);
+long *fillsrc(long *src, int ac, char **av);
+void pushswap(long *la, long *lb, int *b, int size);
+void putinb(long *la, long *lb, int *b, int size);
 
 #endif
